src/embedded/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
transformers>=4.24.0
torch==2.4.1
pillow==10.4.0
opencv-python==4.10.0.84
